add storage logic
