+++ src/main/java/ru/yandex/practicum/filmorate/storage/film/InMemoryFilmStorage.java
@@ -1,0 +1,61 @@
+package ru.yandex.practicum.filmorate.storage.film;
+
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.stereotype.Component;
+import ru.yandex.practicum.filmorate.exception.NotFoundException;
+import ru.yandex.practicum.filmorate.exception.ValidationException;
+import ru.yandex.practicum.filmorate.model.Film;
+
+import java.time.LocalDate;
+import java.util.Collection;
+import java.util.HashMap;
+import java.util.Map;
+
+@Component
+@Slf4j
+public class InMemoryFilmStorage implements FilmStorage {
+
+    private static final LocalDate MIN_RELEASE = LocalDate.of(1895, 12, 28);
+    private final Map<Long, Film> films = new HashMap<>();
+
+    public Collection<Film> getAll() {
+        return films.values();
+    }
+
+    public Film create(Film film) {
+
+        if (film.getReleaseDate().isBefore(MIN_RELEASE)) {
+            throw new ValidationException("Дата релиза должна быть не ранее 28 декабря 1895 года");
+        }
+
+        film.setId(getNextId());
+        films.put(film.getId(), film);
+
+        log.info("Добавлен новый фильм: {}", film);
+
+        return film;
+    }
+
+    public Film update(Film newFilm) {
+
+        if (newFilm.getId() == null) {
+            throw new ValidationException("ID фильма должен быть указан");
+        }
+
+        if (films.containsKey(newFilm.getId())) {
+            films.put(newFilm.getId(), newFilm);
+            log.info("Фильм обновлен: {}", newFilm);
+            return newFilm;
+        }
+        throw new NotFoundException("Пользователь с id = " + newFilm.getId() + " не найден");
+    }
+
+    private long getNextId() {
+        long currentMaxId = films.keySet()
+                .stream()
+                .mapToLong(id -> id)
+                .max()
+                .orElse(0);
+        return ++currentMaxId;
+    }
+}
