+++ src/main/java/ru/yandex/practicum/filmorate/storage/user/InMemoryUserStorage.java
@@ -1,0 +1,69 @@
+package ru.yandex.practicum.filmorate.storage.user;
+
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.stereotype.Component;
+import ru.yandex.practicum.filmorate.exception.NotFoundException;
+import ru.yandex.practicum.filmorate.exception.ValidationException;
+import ru.yandex.practicum.filmorate.model.User;
+
+import java.util.Collection;
+import java.util.HashMap;
+import java.util.Map;
+
+@Component
+@Slf4j
+public class InMemoryUserStorage implements UserStorage {
+
+    private Map<Long, User> users = new HashMap<>();
+
+    public Collection<User> getAll() {
+        return users.values();
+    }
+
+    public User create(User user) {
+
+        if (user.getName() == null) {
+            user.setName(user.getLogin());
+        }
+
+        user.setId(getNextId());
+
+        users.put(user.getId(), user);
+        log.info("Пользователь создан: {}", user);
+        return user;
+    }
+
+    public User update(User newUser) {
+
+        if (newUser.getId() == null) {
+            throw new ValidationException("ID пользователя должен быть указан");
+        }
+
+        if (users.containsKey(newUser.getId())) {
+            User oldUser = users.get(newUser.getId());
+
+            if (newUser.getEmail() != null && users.values().stream()
+                    .anyMatch(user -> user.getEmail().equals(newUser.getEmail()))) {
+                throw new ValidationException("Этот имейл уже используется");
+            } else oldUser.setEmail(newUser.getEmail());
+
+            if (newUser.getName() != null) {
+                oldUser.setName(newUser.getName());
+            }
+            oldUser.setLogin(newUser.getLogin());
+            oldUser.setBirthday(newUser.getBirthday());
+            log.info("Информация о пользователе обновлена: {}", oldUser);
+            return oldUser;
+        }
+        throw new NotFoundException("Пользователь с id = " + newUser.getId() + " не найден");
+    }
+
+    private long getNextId() {
+        long currentMaxId = users.keySet()
+                .stream()
+                .mapToLong(id -> id)
+                .max()
+                .orElse(0);
+        return ++currentMaxId;
+    }
+}
